# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music streaming API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Track {
  id: number;
  title: string;
  albumId?: number;
  albumTitle?: string;
  albumCoverUrl?: string;
  /** Duration in seconds */
  duration: number;
  trackNumber: number;
  audioUrl: string;
  genre?: string;
  releaseYear?: number;
}

export interface Album {
  id: number;
  title: string;
  coverUrl: string;
  releaseYear: number;
  genre?: string;
  trackCount: number;
  totalDuration: number;
}

export interface AlbumWithTracks {
  id: number;
  title: string;
  coverUrl: string;
  releaseYear: number;
  genre?: string;
  tracks: Track[];
}

export type ArtistSocialLinks = {
  instagram?: string;
  twitter?: string;
  youtube?: string;
};

export interface Artist {
  name: string;
  bio: string;
  avatarUrl: string;
  genre?: string;
  location?: string;
  socialLinks?: ArtistSocialLinks;
}

export interface Stats {
  totalTracks: number;
  totalAlbums: number;
  totalListens: number;
}

export type ListTracksParams = {
  albumId?: number;
  search?: string;
};