# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Music streaming API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tracks:
    get:
      operationId: listTracks
      tags: [tracks]
      summary: List all tracks
      parameters:
        - name: albumId
          in: query
          required: false
          schema:
            type: integer
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of tracks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Track"
  /tracks/{id}:
    get:
      operationId: getTrack
      tags: [tracks]
      summary: Get a single track
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Track details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "404":
          description: Not found
...
components:
  schemas:
...
    Track:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        albumId:
          type: integer
        albumTitle:
          type: string
        albumCoverUrl:
          type: string
        duration:
          type: integer
          description: Duration in seconds
        trackNumber:
          type: integer
        audioUrl:
          type: string
        genre:
          type: string
        releaseYear:
          type: integer
      required:
        - id
        - title
        - duration
        - trackNumber
        - audioUrl
...