# lib/api-zod/src/generated/types/track.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music streaming API
 * OpenAPI spec version: 0.1.0
 */

export interface Track {
  id: number;
  title: string;
  albumId?: number;
  albumTitle?: string;
  albumCoverUrl?: string;
  /** Duration in seconds */
  duration: number;
  trackNumber: number;
  audioUrl: string;
  genre?: string;
  releaseYear?: number;
}